Refactor to move FileFormatConfig to the common crate as well
