+++ src/daft-micropartition/src/micropartition.rs
@@ -5,6 +5,7 @@ use std::{ops::Deref, sync::Mutex};
 
 use arrow2::io::parquet::read::schema::infer_schema_with_options;
 use common_error::DaftResult;
+use common_file_formats::{CsvSourceConfig, FileFormatConfig, ParquetSourceConfig};
 use daft_core::prelude::*;
 use daft_csv::{CsvConvertOptions, CsvParseOptions, CsvReadOptions};
 use daft_dsl::ExprRef;
@@ -12,21 +13,20 @@ use daft_json::{JsonConvertOptions, JsonParseOptions, JsonReadOptions};
 use daft_parquet::read::{
     read_parquet_bulk, read_parquet_metadata_bulk, ParquetSchemaInferenceOptions,
 };
-use daft_scan::file_format::{CsvSourceConfig, FileFormatConfig, ParquetSourceConfig};
 use daft_scan::storage_config::{NativeStorageConfig, StorageConfig};
 use daft_scan::{ChunkSpec, DataSource, Pushdowns, ScanTask};
 use daft_table::Table;
 
+use crate::{DaftCSVSnafu, DaftCoreComputeSnafu};
 use parquet2::metadata::FileMetaData;
 use snafu::ResultExt;
-
-#[cfg(feature = "python")]
-use crate::PyIOSnafu;
-use crate::{DaftCSVSnafu, DaftCoreComputeSnafu};
 
 use daft_io::{get_runtime, IOClient, IOConfig, IOStatsContext, IOStatsRef};
 use daft_stats::TableStatistics;
 use daft_stats::{PartitionSpec, TableMetadata};
+
+#[cfg(feature = "python")]
+use {crate::PyIOSnafu, common_file_formats::DatabaseSourceConfig};
 
 #[derive(Debug)]
 pub(crate) enum TableState {
@@ -356,10 +356,7 @@ fn materialize_scan_task(
                     })
                     .collect::<crate::Result<Vec<_>>>()
                 })?,
-                FileFormatConfig::Database(daft_scan::file_format::DatabaseSourceConfig {
-                    sql,
-                    conn,
-                }) => {
+                FileFormatConfig::Database(DatabaseSourceConfig { sql, conn }) => {
                     let predicate = scan_task
                         .pushdowns
                         .filters
