+++ src/daft-scan/src/python.rs
@@ -40,6 +40,8 @@ impl PartialEq for PythonTablesFactoryArgs {
 
 pub mod pylib {
     use common_error::DaftResult;
+    use common_file_formats::python::PyFileFormatConfig;
+    use common_file_formats::FileFormatConfig;
     use common_py_serde::impl_bincode_py_state_serialization;
     use daft_dsl::python::PyExpr;
     use daft_schema::python::field::PyField;
@@ -62,7 +64,6 @@ pub mod pylib {
     use serde::{Deserialize, Serialize};
 
     use crate::anonymous::AnonymousScanOperator;
-    use crate::file_format::FileFormatConfig;
     use crate::storage_config::PythonStorageConfig;
     use crate::DataSource;
     use crate::PartitionField;
@@ -71,7 +72,6 @@ pub mod pylib {
     use crate::ScanOperatorRef;
     use crate::ScanTask;
 
-    use crate::file_format::PyFileFormatConfig;
     use crate::glob::GlobScanOperator;
     use crate::storage_config::PyStorageConfig;
     use common_daft_config::PyDaftExecutionConfig;
